Update : metaData 적용이 필요한 페이지를 server component로 변경
